+++ lib/ua_inspector/downloader/readme.ex
@@ -1,35 +1,9 @@
 defmodule UAInspector.Downloader.README do
   @moduledoc false
 
-  require Logger
-
   alias UAInspector.Config
 
   @readme "ua_inspector.readme.md"
-
-  @doc false
-  @spec path_local() :: Path.t()
-  def path_local do
-    _ =
-      Logger.info(
-        "UAInspector.Downloader.README.path_local/0 has been" <>
-          " declared internal and will eventually be removed."
-      )
-
-    Path.join(Config.database_path(), @readme)
-  end
-
-  @doc false
-  @spec path_priv() :: Path.t()
-  def path_priv do
-    _ =
-      Logger.info(
-        "UAInspector.Downloader.README.path_priv/0 has been" <>
-          " declared internal and will eventually be removed."
-      )
-
-    Application.app_dir(:ua_inspector, ["priv", @readme])
-  end
 
   @doc """
   Writes the informational README file if remote database is the default.